Dirty hack to demonstrate that, with correct cache, it works
I've just forced a call to ->register() from ->process() to
verify that, once that forbiddenFunctionNames is properly updated
the tests (phpcs:set instructions) are observed ok.

This is not the solution, instead... all that code, surely,
should be moved to another method, able to regenerate the cached
information when forbiddenFunctions or patternMatch change.